backend/pyproject.toml
[tool.poetry]
name = "paritet-test"
version = "0.1.0"
description = "Test assignment: Django REST image API"
authors = ["CodeFramer <52332286+CodeFramer0@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
Django = "^4.2"
djangorestframework = "^3.14"
django-filter = "^23.2"
psycopg2 = "^2.9"
environs = "^9.5"
gunicorn = "^21.2"
pillow = "^10.2"
marshmallow = "<3.20.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
